refactor: move peer
