+++ src/services/cluster/actors/actor.rs
@@ -1,9 +1,9 @@
+use super::replication::Replication;
+use super::PeerState;
 use crate::services::cluster::command::AddPeer;
 use crate::services::cluster::command::ClusterCommand;
-use super::peer::Peer;
-use super::replication::Replication;
-use super::types::PeerIdentifier;
-use super::PeerState;
+use crate::services::cluster::peer::identifier::PeerIdentifier;
+use crate::services::cluster::peer::peer::Peer;
 use crate::services::interface::TWrite;
 use crate::services::query_io::QueryIO;
 use std::collections::BTreeMap;
@@ -110,7 +110,7 @@ impl ClusterActor {
     async fn remove_peer(&mut self, peer_addr: PeerIdentifier) -> Option<()> {
         if let Some(peer) = self.members.remove(&peer_addr) {
             // stop the runnin process and take the connection in case topology changes are made
-            let _read_connected = peer.listner_kill_trigger.kill().await;
+            let _read_connected = peer.listener_kill_trigger.kill().await;
             return Some(());
         }
         None
