+++ src/services/cluster/actors/mod.rs
@@ -1,7 +1,4 @@
 pub(super) mod actor;
 pub(crate) use replication::PeerState;
-mod listening_actor;
-
-pub(crate) mod peer;
+pub(crate) mod listening_actor;
 pub(crate) mod replication;
-pub(crate) mod types;
