+++ src/services/cluster/mod.rs
@@ -3,3 +3,4 @@ pub mod inbound;
 pub mod manager;
 pub mod outbound;
 pub mod command;
+pub(crate) mod peer;
