+++ src/services/cluster/peer/identifier.rs
@@ -2,35 +2,6 @@ use std::str::FromStr;
 
 use crate::{from_to, make_smart_pointer};
 
-use super::replication::Replication;
-
-#[derive(Clone, Debug)]
-pub enum PeerKind {
-    Peer,
-    Replica,
-    Master,
-}
-
-impl PeerKind {
-    pub fn accepted_peer_kind(self_repl_id: &str, other_repl_id: &str) -> Self {
-        match other_repl_id {
-            "?" => Self::Replica,
-            id if id == self_repl_id => Self::Replica,
-            _ => Self::Peer,
-        }
-    }
-    pub fn connected_peer_kind(self_repl_info: &Replication, other_repl_id: &str) -> Self {
-        if self_repl_info.master_replid == "?" {
-            Self::Master
-        } else if self_repl_info.master_replid == other_repl_id {
-            Self::Replica
-        } else {
-            Self::Peer
-        }
-    }
-}
-
-// TODO refactor the following so it allows for host and port to be separate
 #[derive(Debug, PartialEq, Eq, PartialOrd, Ord, Clone, Hash)]
 pub struct PeerIdentifier(pub String);
 impl PeerIdentifier {
