+++ src/services/cluster/peer/peer.rs
@@ -1,17 +1,16 @@
-use super::types::PeerIdentifier;
-use crate::services::cluster::command::ClusterCommand;
 use crate::services::cluster::actors::listening_actor::ListeningActorKillTrigger;
 use crate::services::cluster::actors::listening_actor::PeerListeningActor;
-use crate::services::cluster::actors::types::PeerKind;
-use tokio::net::tcp::OwnedReadHalf;
-use tokio::net::tcp::OwnedWriteHalf;
+use crate::services::cluster::command::ClusterCommand;
+use crate::services::cluster::peer::connected_types::WriteConnected;
+use crate::services::cluster::peer::identifier::PeerIdentifier;
+use crate::services::cluster::peer::kind::PeerKind;
 use tokio::net::TcpStream;
 use tokio::sync::mpsc::Sender;
 
 #[derive(Debug)]
 pub(crate) struct Peer {
     pub(crate) w_conn: WriteConnected,
-    pub(crate) listner_kill_trigger: ListeningActorKillTrigger,
+    pub(crate) listener_kill_trigger: ListeningActorKillTrigger,
     pub(crate) last_seen: std::time::Instant,
 }
 
@@ -35,32 +34,9 @@ impl Peer {
 
         Self {
             w_conn: write_connected,
-            listner_kill_trigger: ListeningActorKillTrigger::new(kill_trigger, listening_task),
+            listener_kill_trigger: ListeningActorKillTrigger::new(kill_trigger, listening_task),
             last_seen: std::time::Instant::now(),
         }
     }
 }
 
-#[derive(Debug)]
-pub struct WriteConnected {
-    pub stream: OwnedWriteHalf,
-    pub kind: PeerKind,
-}
-
-#[derive(Debug)]
-pub(super) struct ReadConnected {
-    pub stream: OwnedReadHalf,
-    pub kind: PeerKind,
-}
-
-impl From<(OwnedReadHalf, PeerKind)> for ReadConnected {
-    fn from((r, peer_kind): (OwnedReadHalf, PeerKind)) -> ReadConnected {
-        ReadConnected { stream: r, kind: peer_kind }
-    }
-}
-
-impl From<(OwnedWriteHalf, PeerKind)> for WriteConnected {
-    fn from((w, peer_kind): (OwnedWriteHalf, PeerKind)) -> WriteConnected {
-        WriteConnected { stream: w, kind: peer_kind }
-    }
-}
